refactor: Move exercise set components to dedicated folder
Move `SwipableDayView`, `ExerciseSet` and related components into a new `exerciseSet` folder to improve project organization.  This improves maintainability and readability of the project's structure.
