+++ frontend/src/components/dashboard/exerciseSet/SwipableDayView.tsx
